`dbmock` folder moved from `smeup` folder to root, for using it both in `smeup` and `db`
