+++ rpgJavaInterpreter-core/src/test/kotlin/com/smeup/rpgparser/dbmock/C5ADFF9LDbMock.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.smeup.rpgparser.smeup.dbmock
+package com.smeup.rpgparser.dbmock
 
 import com.smeup.rpgparser.interpreter.FileMetadata
 import com.smeup.rpgparser.utils.DbMock
