+++ rpgJavaInterpreter-core/src/test/kotlin/com/smeup/rpgparser/dbmock/C5RREGHLDbMock.kt
@@ -1,4 +1,4 @@
-package com.smeup.rpgparser.smeup.dbmock
+package com.smeup.rpgparser.dbmock
 
 import com.smeup.rpgparser.interpreter.FileMetadata
 import com.smeup.rpgparser.utils.DbMock
