+++ rpgJavaInterpreter-core/src/test/kotlin/com/smeup/rpgparser/smeup/MULANGT10BaseCodopTest.kt
@@ -1,6 +1,6 @@
 package com.smeup.rpgparser.smeup
 
-import com.smeup.rpgparser.smeup.dbmock.C5ADFF9LDbMock
+import com.smeup.rpgparser.dbmock.C5ADFF9LDbMock
 import org.junit.Test
 import kotlin.test.assertEquals
